+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [1.0.1-next.6](https://github.com/warp-ds/react/compare/v1.0.1-next.5...v1.0.1-next.6) (2023-09-20)
+
+
+### Bug Fixes
+
+* Add fullWidth button prop ([#89](https://github.com/warp-ds/react/issues/89)) ([4cdbe5d](https://github.com/warp-ds/react/commit/4cdbe5d3c42f21c938ac5c416824c677601fb8ad))
+
+## [1.0.1-next.5](https://github.com/warp-ds/react/compare/v1.0.1-next.4...v1.0.1-next.5) (2023-09-19)
+
+
+### Bug Fixes
+
+* pass onClick to anchor in button component ([#87](https://github.com/warp-ds/react/issues/87)) ([24a4247](https://github.com/warp-ds/react/commit/24a4247fe3c58db07c69ed9e56ea00d825e575ed))
+
+## [1.0.1-next.4](https://github.com/warp-ds/react/compare/v1.0.1-next.3...v1.0.1-next.4) (2023-09-18)
+
+
+### Bug Fixes
+
+* **button:** set secondary variant styles by default ([#86](https://github.com/warp-ds/react/issues/86)) ([cb0c60f](https://github.com/warp-ds/react/commit/cb0c60f7df0711fa901a09fb28b9126682ebd387))
+
 ## [1.0.1-next.3](https://github.com/warp-ds/react/compare/v1.0.1-next.2...v1.0.1-next.3) (2023-09-07)